requirements.txt
openai==1.51.0
praw==7.7.1
tweepy==4.14.0
python-dotenv==1.0.1
